changed color bar models to support categories incl. labels
