+++ src/components/ColorBarLegend/ColorBarLegend.tsx
@@ -22,165 +22,14 @@
  * SOFTWARE.
  */
 
-import React, { useState } from "react";
-import makeStyles from "@mui/styles/makeStyles";
-import { Theme } from "@mui/material";
-import Popover from "@mui/material/Popover";
+import { ColorBarLegendProps } from "./common";
+import ColorBarLegendCategorical from "./ColorBarLegendCategorical";
+import ColorBarLegendContinuous from "./ColorBarLegendContinuous";
 
-import { ColorBar, ColorBars } from "@/model/colorBar";
-import { UserColorBar } from "@/model/userColorBar";
-import ColorBarCanvas from "./ColorBarCanvas";
-import ColorBarColorEditor from "./ColorBarColorEditor";
-import ColorBarRangeEditor from "./ColorBarRangeEditor";
-import ColorBarLabels from "./ColorBarLabels";
-
-const useStyles = makeStyles((theme: Theme) => ({
-  title: {
-    fontSize: "x-small",
-    fontWeight: "bold",
-    width: "100%",
-    display: "flex",
-    flexWrap: "nowrap",
-    justifyContent: "center",
-    paddingBottom: theme.spacing(0.5),
-  },
-  container: {
-    paddingLeft: theme.spacing(1.5),
-    paddingRight: theme.spacing(1.5),
-    paddingBottom: theme.spacing(0.5),
-    paddingTop: theme.spacing(0.5),
-    color: "black",
-  },
-}));
-
-interface ColorBarLegendProps {
-  variableName: string | null;
-  variableUnits: string;
-  variableColorBarMinMax: [number, number];
-  variableColorBarName: string;
-  variableColorBar: ColorBar;
-  variableOpacity: number;
-  updateVariableColorBar: (
-    colorBarMinMax: [number, number],
-    colorBarName: string,
-    opacity: number,
-  ) => void;
-  colorBars: ColorBars;
-  userColorBars: UserColorBar[];
-  addUserColorBar: (userColorBarId: string) => void;
-  removeUserColorBar: (userColorBarId: string) => void;
-  updateUserColorBar: (userColorBar: UserColorBar) => void;
-  updateUserColorBars: (userColorBars: UserColorBar[]) => void;
-  width?: number | string;
-  height?: number | string;
-  numTicks?: number;
-}
-
-export default function ColorBarLegend({
-  variableName,
-  variableUnits,
-  variableColorBarMinMax,
-  variableColorBarName,
-  variableColorBar,
-  variableOpacity,
-  updateVariableColorBar,
-  colorBars,
-  userColorBars,
-  addUserColorBar,
-  removeUserColorBar,
-  updateUserColorBar,
-  updateUserColorBars,
-  width,
-  height,
-  numTicks,
-}: ColorBarLegendProps) {
-  const classes = useStyles();
-
-  const [colorBarRangeEditorAnchor, setColorBarRangeEditorAnchor] =
-    useState<HTMLDivElement | null>(null);
-  const [colorBarSelectAnchor, setColorBarSelectAnchor] =
-    useState<HTMLCanvasElement | null>(null);
-
-  if (!variableName) {
-    return null;
-  }
-
-  const handleOpenColorBarRangeEditor = (
-    event: React.MouseEvent<HTMLDivElement>,
-  ) => {
-    setColorBarRangeEditorAnchor(event.currentTarget);
-  };
-
-  const handleCloseColorBarRangeEditor = () => {
-    setColorBarRangeEditorAnchor(null);
-  };
-
-  const handleOpenColorBarSelect = (
-    event: React.MouseEvent<HTMLCanvasElement>,
-  ) => {
-    setColorBarSelectAnchor(event.currentTarget);
-  };
-
-  const handleCloseColorBarSelect = () => {
-    setColorBarSelectAnchor(null);
-  };
-
-  const variableTitle = `${variableName} (${variableUnits || "-"})`;
-
-  return (
-    <div className={"ol-control " + classes.container}>
-      <div className={classes.title}>
-        <span>{variableTitle}</span>
-      </div>
-      <ColorBarCanvas
-        colorBar={variableColorBar}
-        opacity={variableOpacity}
-        width={width}
-        height={height}
-        onClick={handleOpenColorBarSelect}
-      />
-      <ColorBarLabels
-        minValue={variableColorBarMinMax[0]}
-        maxValue={variableColorBarMinMax[1]}
-        numTicks={numTicks || 5}
-        onClick={handleOpenColorBarRangeEditor}
-      />
-      <Popover
-        anchorEl={colorBarRangeEditorAnchor}
-        open={Boolean(colorBarRangeEditorAnchor)}
-        onClose={handleCloseColorBarRangeEditor}
-        anchorOrigin={{ vertical: "bottom", horizontal: "left" }}
-        transformOrigin={{ vertical: "top", horizontal: "center" }}
-      >
-        <ColorBarRangeEditor
-          variableTitle={variableTitle}
-          variableColorBarMinMax={variableColorBarMinMax}
-          variableColorBarName={variableColorBarName}
-          variableOpacity={variableOpacity}
-          updateVariableColorBar={updateVariableColorBar}
-        />
-      </Popover>
-      <Popover
-        anchorEl={colorBarSelectAnchor}
-        open={Boolean(colorBarSelectAnchor)}
-        onClose={handleCloseColorBarSelect}
-        anchorOrigin={{ vertical: "bottom", horizontal: "left" }}
-        transformOrigin={{ vertical: "top", horizontal: "left" }}
-      >
-        <ColorBarColorEditor
-          variableColorBarMinMax={variableColorBarMinMax}
-          variableColorBarName={variableColorBarName}
-          variableColorBar={variableColorBar}
-          variableOpacity={variableOpacity}
-          updateVariableColorBar={updateVariableColorBar}
-          colorBars={colorBars}
-          userColorBars={userColorBars}
-          addUserColorBar={addUserColorBar}
-          removeUserColorBar={removeUserColorBar}
-          updateUserColorBar={updateUserColorBar}
-          updateUserColorBars={updateUserColorBars}
-        />
-      </Popover>
-    </div>
+export default function ColorBarLegend(props: ColorBarLegendProps) {
+  return props.variableColorBar.categories ? (
+    <ColorBarLegendCategorical {...props} />
+  ) : (
+    <ColorBarLegendContinuous {...props} />
   );
 }
