+++ src/components/ColorBarLegend/common.ts
@@ -1,0 +1,71 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2019-2024 by the xcube development team and contributors.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to
+ * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
+ * of the Software, and to permit persons to whom the Software is furnished to do
+ * so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+import makeStyles from "@mui/styles/makeStyles";
+import { Theme } from "@mui/material";
+
+import { ColorBar, ColorBars } from "@/model/colorBar";
+import { UserColorBar } from "@/model/userColorBar";
+
+export const useColorBarLegendStyles = makeStyles((theme: Theme) => ({
+  title: {
+    fontSize: "x-small",
+    fontWeight: "bold",
+    width: "100%",
+    display: "flex",
+    flexWrap: "nowrap",
+    justifyContent: "center",
+    paddingBottom: theme.spacing(0.5),
+  },
+  container: {
+    paddingLeft: theme.spacing(1.5),
+    paddingRight: theme.spacing(1.5),
+    paddingBottom: theme.spacing(0.5),
+    paddingTop: theme.spacing(0.5),
+    color: "black",
+  },
+}));
+
+export interface ColorBarLegendProps {
+  variableName: string | null;
+  variableUnits: string;
+  variableColorBarMinMax: [number, number];
+  variableColorBarName: string;
+  variableColorBar: ColorBar;
+  variableOpacity: number;
+  updateVariableColorBar: (
+    colorBarMinMax: [number, number],
+    colorBarName: string,
+    opacity: number,
+  ) => void;
+  colorBars: ColorBars;
+  userColorBars: UserColorBar[];
+  addUserColorBar: (userColorBarId: string) => void;
+  removeUserColorBar: (userColorBarId: string) => void;
+  updateUserColorBar: (userColorBar: UserColorBar) => void;
+  updateUserColorBars: (userColorBars: UserColorBar[]) => void;
+  width?: number | string;
+  height?: number | string;
+  numTicks?: number;
+}
